-feat Create MapSystem class to store map-related methods
